requirements.txt
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
jiter==0.4.1
numpy==1.26.4
openai==1.30.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1